Polish menu before creating platform window
A style may want to change the surface format of the window. Changing the
surface format after the window has been created has no effect though.
One example of this is Breeze where the context menus created here had an
ugly border instead of the intended look. This is because the setting of
Qt::WA_TranslucentBackground by breeze had no effect anymore after the
window was already created.